Move cardano specific keypair to `modifications` module
