+++ rust/tw_keypair/src/ed25519/modifications/cardano/extended_public.rs
