+++ rust/tw_keypair/src/ed25519/modifications/cardano/mod.rs
@@ -1,0 +1,13 @@
+// Copyright © 2017-2023 Trust Wallet.
+//
+// This file is part of Trust. The full Trust copyright notice, including
+// terms governing use, modification, and redistribution, is contained in the
+// file LICENSE at the root of the source code distribution tree.
+
+use sha2::Sha512;
+
+pub(crate) mod extended_private;
+pub(crate) mod extended_public;
+
+pub type ExtendedPrivateKey = extended_private::ExtendedPrivateKey<Sha512>;
+pub type ExtendedPublicKey = extended_public::ExtendedPublicKey<Sha512>;
