+++ rust/tw_keypair/src/ed25519/modifications/mod.rs
@@ -1,0 +1,7 @@
+// Copyright © 2017-2023 Trust Wallet.
+//
+// This file is part of Trust. The full Trust copyright notice, including
+// terms governing use, modification, and redistribution, is contained in the
+// file LICENSE at the root of the source code distribution tree.
+
+pub mod cardano;
